refactor: CertUtil SPI
