+++ .github/workflows/deploy.yml
@@ -3,8 +3,8 @@ name: deploy
 on:
   workflow_dispatch:
   push:
-    branches:
-      - master
+    tag:
+      - "v*"
 jobs:
   deploy:
     runs-on: ubuntu-latest
@@ -21,15 +21,10 @@ jobs:
         with: # running setup-java again overwrites the settings.xml
           java-version: 1.8
           server-id: releases # Value of the distributionManagement/repository/id field of the pom.xml
-          server-username: monkeyWie # env variable for username in deploy
-          server-password: ${{ secrets.MAVEN_CENTRAL_TOKEN }} # env variable for token in deploy
+          server-username: MAVEN_USERNAME # env variable for username in deploy
+          server-password: MAVEN_CENTRAL_TOKEN # env variable for token in deploy
+          gpg-passphrase: MAVEN_GPG_PASSPHRASE # env variable for GPG private key passphrase
           gpg-private-key: ${{ secrets.MAVEN_GPG_PRIVATE_KEY }} # Value of the GPG private key to import
-          gpg-passphrase: ${{ secrets.MAVEN_GPG_PASSPHRASE }} # env variable for GPG private key passphrase
-
-      - name: Debug gpg
-        run: |
-          cat /home/runner/.m2/settings.xml
-          gpg --list-secret-keys --keyid-format LONG
 
       - name: Publish to Apache Maven Central
         run: mvn clean deploy -P release
