+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>com.github.monkeywie</groupId>
     <artifactId>proxyee</artifactId>
-    <version>1.3.0</version>
+    <version>1.3.1</version>
     <build>
         <plugins>
             <plugin>
