+++ src/main/java/com/github/monkeywie/proxyee/crt/service/bc/BouncyCastleCertGenerator.java
@@ -1,5 +1,7 @@
-package com.github.monkeywie.proxyee.crt;
+package com.github.monkeywie.proxyee.crt.service.bc;
 
+import com.github.monkeywie.proxyee.crt.spi.CertGenerator;
+import com.github.monkeywie.proxyee.crt.spi.CertGeneratorInfo;
 import org.bouncycastle.asn1.x500.X500Name;
 import org.bouncycastle.asn1.x509.BasicConstraints;
 import org.bouncycastle.asn1.x509.Extension;
