--- src/main/resources/META-INF/services/com.github.monkeywie.proxyee.crt.CertGenerator
@@ -1,1 +1,0 @@
-com.github.monkeywie.proxyee.crt.BouncyCastleCertGenerator
