+++ src/main/resources/META-INF/services/com.github.monkeywie.proxyee.crt.spi.CertGenerator
@@ -1,0 +1,1 @@
+com.github.monkeywie.proxyee.crt.service.bc.BouncyCastleCertGenerator
